// frontend/src/app/model/driver-guide-package.ts
export class DriverGuidePackage {
    id: string;
    account: string;
    name: string;
    description: string;
    price: number;
    noOfPeoples: number;
    insertUserId: string;
    updateUserId: string;
    insertDateTime: Date;
    updateDateTime: Date;
    status: number;
    sortOrder: number;
}